Refactor JS to purposeful files
