+++ portfolio/src/main/webapp/index.html
@@ -7,7 +7,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Lato&family=Roboto&display=swap" 
      rel="stylesheet">
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script src="comments.js" type="module"></script>
   </head>
   <body>
     <div id="nav-bar">
