+++ portfolio/src/main/webapp/random-fact.html
@@ -6,7 +6,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Lato&family=Roboto&display=swap" 
      rel="stylesheet">
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script src="random-fact.js" type="module"></script>
   </head>
   <body>
     <div id="nav-bar">
